refactor: extract closed-day override modal into dedicated component and hook
Moves the open-closed-day modal UI, styles, and mutation logic out of
AvailabilityWeekPage into OpenClosedDayModal and useClosedDayOverride,
reducing the page component by ~340 lines.
