+++ src/pages/availability/week/AvailabilityWeekPage.tsx
@@ -1,12 +1,7 @@
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useParams, useSearchParams } from 'react-router-dom';
-import { TextField } from '@mui/material';
-import {
-	createAvailabilityDateOverride,
-	getAppointmentDetailsBySlotId,
-} from '@psycron/api/user/availability';
-import type { AvailabilityDateOverrideMode } from '@psycron/api/user/availability/index.types';
+import { getAppointmentDetailsBySlotId } from '@psycron/api/user/availability';
 import { AvailabilityLegend } from '@psycron/components/availability/AvailabilityLegend';
 import { NavButton } from '@psycron/components/availability/AvailabilityNavButton';
 import { AvailabilityTodayButton } from '@psycron/components/availability/AvailabilityTodayButton';
@@ -19,32 +14,23 @@ import {
 	Filter,
 	FilterFull,
 } from '@psycron/components/icons';
-import { Modal } from '@psycron/components/modal/Modal';
-import { useAlert } from '@psycron/context/alert/AlertContext';
 import { useJupiterAvailabilityConfig } from '@psycron/hooks/useJupiterAvailabilityConfig';
 import { useTherapistId } from '@psycron/hooks/useTherapistId';
 import useViewport from '@psycron/hooks/useViewport';
 import { PageLayout } from '@psycron/layouts/app/pages-layout/PageLayout';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { useQueryClient } from '@tanstack/react-query';
 import { format, isPast, isToday, parseISO } from 'date-fns';
 
 import { DayHeaderPopover } from './day-header-popover/DayHeaderPopover';
 import { AvailabilityWeekDrawer } from './drawer/AvailabilityWeekDrawer';
 import { AvailabilityWeekFilters } from './filters/AvailabilityWeekFilters';
 import { useAvailabilityWeekViewModel } from './hook/useAvailabilityWeekViewModel';
+import { useClosedDayOverride } from './hook/useClosedDayOverride';
 import { useBlockDay, useUnblockDay } from './hook/useDayActions';
+import { OpenClosedDayModal } from './open-closed-day-modal/OpenClosedDayModal';
 import { AvailabilityWeekDesktopGrid } from './views/desktop-view/AvailabilityWeekDesktopGrid';
 import { AvailabilityWeekMobileList } from './views/mobile-view/AvailabilityWeekMobileList';
 import {
-	ClosedDayModalBody,
-	ClosedDayModalText,
-	ClosedDayOptionButton,
-	ClosedDayOptionDescription,
-	ClosedDayOptionsGrid,
-	ClosedDayOptionTitle,
-	ClosedDaySlotButton,
-	ClosedDaySlotsGrid,
-	ClosedDayTimeRangeRow,
 	FilterButton,
 	WeekCard,
 	WeekFeaturesActions,
@@ -58,11 +44,6 @@ import {
 	WeekTitleBlock,
 } from './AvailabilityWeekPage.styles';
 import type { IWeekSlot } from './AvailabilityWeekPage.types';
-import {
-	generateSlotStartTimes,
-	parseDurationMinutes,
-	parseTimeRange,
-} from './AvailabilityWeekPage.utils';
 
 const parseDebugNowMinutes = (value: string | null): number | null => {
 	if (!value || !import.meta.env.DEV) return null;
@@ -93,7 +74,6 @@ export const AvailabilityWeekPage = () => {
 	const { date } = useParams<{ date: string }>();
 	const [searchParams] = useSearchParams();
 	const { isMobile } = useViewport();
-	const { showAlert } = useAlert();
 	const {
 		activeFilterCount,
 		allSessionTypes,
@@ -128,21 +108,15 @@ export const AvailabilityWeekPage = () => {
 		null
 	);
 	const [dayPopoverDate, setDayPopoverDate] = useState<string | null>(null);
-	const [defaultBlockedDate, setDefaultBlockedDate] = useState<string | null>(
-		null
-	);
-	const [defaultBlockedMode, setDefaultBlockedMode] =
-		useState<AvailabilityDateOverrideMode>('FULL_DAY');
-	const [overrideStartTime, setOverrideStartTime] = useState('09:00');
-	const [overrideEndTime, setOverrideEndTime] = useState('17:00');
-	const [selectedSpecificSlots, setSelectedSpecificSlots] = useState<string[]>(
-		[]
-	);
 	const [shouldScrollToToday, setShouldScrollToToday] = useState(false);
 	const debugNowMinutes = parseDebugNowMinutes(searchParams.get('debugNow'));
 
 	const queryClient = useQueryClient();
 	const therapistId = useTherapistId();
+	const closedDayOverride = useClosedDayOverride({
+		availability,
+		therapistId,
+	});
 
 	const blockDay = useBlockDay(therapistId, () => {
 		setDayPopoverDate(null);
@@ -150,124 +124,18 @@ export const AvailabilityWeekPage = () => {
 	const unblockDay = useUnblockDay(therapistId, () => {
 		setDayPopoverDate(null);
 	});
-	const openClosedDay = useMutation({
-		mutationFn: async () => {
-			if (!defaultBlockedDate) {
-				throw new Error('Missing override date');
-			}
-
-			if (defaultBlockedMode === 'FULL_DAY') {
-				return createAvailabilityDateOverride({
-					date: defaultBlockedDate,
-					mode: 'FULL_DAY',
-					therapistId,
-				});
-			}
-
-			if (defaultBlockedMode === 'TIME_RANGE') {
-				return createAvailabilityDateOverride({
-					date: defaultBlockedDate,
-					endTime: overrideEndTime,
-					mode: 'TIME_RANGE',
-					startTime: overrideStartTime,
-					therapistId,
-				});
-			}
-
-			return createAvailabilityDateOverride({
-				date: defaultBlockedDate,
-				mode: 'SPECIFIC_SLOTS',
-				slotStartTimes: selectedSpecificSlots,
-				therapistId,
-			});
-		},
-		onError: () => {
-			showAlert({
-				message: t('availability.week.default-blocked-day.error'),
-				severity: 'error',
-			});
-		},
-		onSuccess: () => {
-			queryClient.invalidateQueries({ queryKey: ['therapistAvailability'] });
-			showAlert({
-				message: t('availability.week.default-blocked-day.success'),
-				severity: 'success',
-			});
-			setDefaultBlockedDate(null);
-			setSelectedSpecificSlots([]);
-		},
-	});
-
-	const baseTimeRange = useMemo(() => {
-		if (!availability?.timeRange) return null;
-		try {
-			return parseTimeRange(availability.timeRange);
-		} catch {
-			return null;
-		}
-	}, [availability?.timeRange]);
-
-	const sessionDurationMinutes = useMemo(() => {
-		if (!availability?.sessionDuration) return null;
-		try {
-			return parseDurationMinutes(availability.sessionDuration);
-		} catch {
-			return null;
-		}
-	}, [availability?.sessionDuration]);
-
-	useEffect(() => {
-		if (!defaultBlockedDate || !baseTimeRange) return;
-
-		setDefaultBlockedMode('FULL_DAY');
-		setOverrideStartTime(baseTimeRange.startTime);
-		setOverrideEndTime(baseTimeRange.endTime);
-		setSelectedSpecificSlots([]);
-	}, [baseTimeRange, defaultBlockedDate]);
-
-	const specificSlotOptions =
-		baseTimeRange && sessionDurationMinutes
-			? generateSlotStartTimes(
-					baseTimeRange.startTime,
-					baseTimeRange.endTime,
-					sessionDurationMinutes
-				)
-			: [];
-
-	const partialSlotOptions =
-		sessionDurationMinutes && overrideStartTime < overrideEndTime
-			? generateSlotStartTimes(
-					overrideStartTime,
-					overrideEndTime,
-					sessionDurationMinutes
-				)
-			: [];
-
-	const isClosedDayConfirmDisabled =
-		!availability?.timeRange ||
-		!availability?.sessionDuration ||
-		(defaultBlockedMode === 'TIME_RANGE' && partialSlotOptions.length === 0) ||
-		(defaultBlockedMode === 'SPECIFIC_SLOTS' && selectedSpecificSlots.length === 0);
 
 	const handleDayHeaderClick = (dateStr: string) => {
 		const dayDate = parseISO(dateStr);
 		if (isPast(dayDate) && !isToday(dayDate)) return;
 		if (getDaySlots(dayDate).length === 0) {
-			setDefaultBlockedDate(dateStr);
+			closedDayOverride.open(dateStr);
 			return;
 		}
 		setDayPopoverDate(dateStr);
 	};
 
 	const handleSlotClick = (slot: IWeekSlot) => setSelectedSlot(slot);
-
-	const handleSpecificSlotToggle = (startTime: string) => {
-		setSelectedSpecificSlots((current) =>
-			current.includes(startTime)
-				? current.filter((item) => item !== startTime)
-				: [...current, startTime].sort()
-		);
-	};
 
 	const handleTodayClick = useCallback(() => {
 		setShouldScrollToToday(true);
@@ -469,113 +337,23 @@ export const AvailabilityWeekPage = () => {
 					);
 				})()}
 
-			{defaultBlockedDate && (
-				<Modal
-					openModal
-					title={t('availability.week.default-blocked-day.title')}
-					onClose={() => setDefaultBlockedDate(null)}
-					cardActionsProps={{
-						actionName: t('availability.week.default-blocked-day.confirm'),
-						disabled: isClosedDayConfirmDisabled,
-						hasSecondAction: true,
-						loading: openClosedDay.isPending,
-						onClick: () => openClosedDay.mutate(),
-						secondAction: () => setDefaultBlockedDate(null),
-						secondActionName: t('availability.week.drawer.cancel-back'),
-					}}
-				>
-					<ClosedDayModalBody>
-						<ClosedDayModalText>
-							{t('availability.week.default-blocked-day.body', {
-								day: format(parseISO(defaultBlockedDate), 'EEEE, MMMM d'),
-							})}
-						</ClosedDayModalText>
-
-						<ClosedDayOptionsGrid>
-							<ClosedDayOptionButton
-								isSelected={defaultBlockedMode === 'FULL_DAY'}
-								onClick={() => setDefaultBlockedMode('FULL_DAY')}
-								type='button'
-							>
-								<ClosedDayOptionTitle>
-									{t('availability.week.default-blocked-day.open-full-day')}
-								</ClosedDayOptionTitle>
-								<ClosedDayOptionDescription>
-									{t(
-										'availability.week.default-blocked-day.open-full-day-desc'
-									)}
-								</ClosedDayOptionDescription>
-							</ClosedDayOptionButton>
-							<ClosedDayOptionButton
-								isSelected={defaultBlockedMode === 'TIME_RANGE'}
-								onClick={() => setDefaultBlockedMode('TIME_RANGE')}
-								type='button'
-							>
-								<ClosedDayOptionTitle>
-									{t('availability.week.default-blocked-day.open-part-day')}
-								</ClosedDayOptionTitle>
-								<ClosedDayOptionDescription>
-									{t(
-										'availability.week.default-blocked-day.open-part-day-desc'
-									)}
-								</ClosedDayOptionDescription>
-							</ClosedDayOptionButton>
-							<ClosedDayOptionButton
-								isSelected={defaultBlockedMode === 'SPECIFIC_SLOTS'}
-								onClick={() => setDefaultBlockedMode('SPECIFIC_SLOTS')}
-								type='button'
-							>
-								<ClosedDayOptionTitle>
-									{t(
-										'availability.week.default-blocked-day.open-specific-slots'
-									)}
-								</ClosedDayOptionTitle>
-								<ClosedDayOptionDescription>
-									{t(
-										'availability.week.default-blocked-day.open-specific-slots-desc'
-									)}
-								</ClosedDayOptionDescription>
-							</ClosedDayOptionButton>
-						</ClosedDayOptionsGrid>
-
-						{defaultBlockedMode === 'TIME_RANGE' && (
-							<ClosedDayTimeRangeRow>
-								<TextField
-									fullWidth
-									label={t(
-										'availability.week.default-blocked-day.start-time'
-									)}
-									onChange={(e) => setOverrideStartTime(e.target.value)}
-									type='time'
-									value={overrideStartTime}
-								/>
-								<TextField
-									fullWidth
-									label={t('availability.week.default-blocked-day.end-time')}
-									onChange={(e) => setOverrideEndTime(e.target.value)}
-									type='time'
-									value={overrideEndTime}
-								/>
-							</ClosedDayTimeRangeRow>
-						)}
-
-						{defaultBlockedMode === 'SPECIFIC_SLOTS' && (
-							<ClosedDaySlotsGrid>
-								{specificSlotOptions.map((slotStartTime) => (
-									<ClosedDaySlotButton
-										isSelected={selectedSpecificSlots.includes(slotStartTime)}
-										key={slotStartTime}
-										onClick={() => handleSpecificSlotToggle(slotStartTime)}
-										type='button'
-									>
-										{slotStartTime}
-									</ClosedDaySlotButton>
-								))}
-							</ClosedDaySlotsGrid>
-						)}
-					</ClosedDayModalBody>
-				</Modal>
-			)}
+			<OpenClosedDayModal
+				endTime={closedDayOverride.endTime}
+				isConfirmDisabled={closedDayOverride.isConfirmDisabled}
+				isLoading={closedDayOverride.isPending}
+				mode={closedDayOverride.mode}
+				openDate={closedDayOverride.overrideDate}
+				overrideSlotOptions={closedDayOverride.overrideSlotOptions}
+				partialSlotOptions={closedDayOverride.partialSlotOptions}
+				selectedSpecificSlots={closedDayOverride.selectedSpecificSlots}
+				startTime={closedDayOverride.startTime}
+				onClose={closedDayOverride.close}
+				onConfirm={closedDayOverride.confirm}
+				onEndTimeChange={closedDayOverride.setEndTime}
+				onModeChange={closedDayOverride.setMode}
+				onSpecificSlotToggle={closedDayOverride.toggleSpecificSlot}
+				onStartTimeChange={closedDayOverride.setStartTime}
+			/>
 
 			{selectedSlot && (
 				<AvailabilityWeekDrawer
